refactor: 요구사항에 따라 패키지명을 ui -> view 로 변경하고 클래스명에서 변경된 패키지명과 중복되는 'view' 를 제거함
