+++ src/main/java/race/view/Result.java
@@ -1,8 +1,8 @@
-package race.ui;
+package race.view;
 
 import java.util.List;
 
-public final class ResultView {
+public final class Result {
 
     private static final String QUESTION_FOR_CAR_NAME = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).";
     public static final String RESULT_SENTENCE = "실행 결과";
@@ -10,7 +10,7 @@ public final class ResultView {
     public static final String BAR = "-";
     private static final String TEXT_FOR_WINNER = "가 최종 우승했습니다.";
 
-    private ResultView() {};
+    private Result() {};
 
     public static void getFirstQuestion() {
         System.out.println(QUESTION_FOR_CAR_NAME);
